不要な記述を削除
